fix: Return useDeleteUser to ready state after error
Reason:
This change is implemented to correct an issue where 'useDeleteUser' would remain in the 'loading' state if an error occurred. Instead, it should return to the 'ready' state.

Changes:
Modified 'useDeleteUser.ts'. Added a try-catch block to detect if a 'FirebaseError' occurred during the deletion of the user. If such an error arises, the state is set to 'ready' and the error is thrown.