fix: Set state to ready in case of sign-in error
To ensure that user interface does not hang or freeze when encountering an error during different signIn options, the state is now set to 'ready' when a 'catch' block is triggered.

Changes:
Added 'try-catch' block in SignIn method for different signIn types such as 'classic', 'link', 'google', 'facebook', 'apple', 'twitter', 'github', 'microsoft', 'yahoo' to handle potential errors. If an error is thrown, it sets the state to 'ready' before re-throwing the error.