fix: Set state to ready on sign up error
Why: To assure that the state is set back to 'ready' even when an error occurs during the sign up process, preventing the state from being stuck at 'loading'.

Changes:
Added a try/catch block around createUserWithEmailAndPassword in useSignUp.ts in order to catch any errors and set the state back to 'ready'.