Fix Auth States on Error (#90)
* fix: Return useDeleteUser to ready state after error

Reason:
This change is implemented to correct an issue where 'useDeleteUser' would remain in the 'loading' state if an error occurred. Instead, it should return to the 'ready' state.

Changes:
Modified 'useDeleteUser.ts'. Added a try-catch block to detect if a 'FirebaseError' occurred during the deletion of the user. If such an error arises, the state is set to 'ready' and the error is thrown.

* fix: Set state to ready in case of sign-in error

To ensure that user interface does not hang or freeze when encountering an error during different signIn options, the state is now set to 'ready' when a 'catch' block is triggered.

Changes:
Added 'try-catch' block in SignIn method for different signIn types such as 'classic', 'link', 'google', 'facebook', 'apple', 'twitter', 'github', 'microsoft', 'yahoo' to handle potential errors. If an error is thrown, it sets the state to 'ready' before re-throwing the error.

* fix: set state to ready in case of an error

why:
To prevent the application from being stuck in 'loading' state when there is an error while sending email verification.

Changes:
1. Introduced a try-catch block in 'useSendEmailVerification'.
2. In case of an error, the state is now set to 'ready' instead of being stuck at 'loading'.
3. The error is then rethrown to be handled by the calling function.

* remove FirebaseError check

* fix: Set state to 'ready' in case of error during sign out

Why: There was a lack of error handling in our sign out operation. If an error occurred during sign out, the state was not being reset back to 'ready', leaving the system in a 'loading' state indefinitely.

Changes:
A try catch block was added to handle any errors occurring during the sign out process. If an error is caught, the state is now set to 'ready' before the error is re-thrown.

* fix: Set state to ready on sign up error

Why: To assure that the state is set back to 'ready' even when an error occurs during the sign up process, preventing the state from being stuck at 'loading'.

Changes:
Added a try/catch block around createUserWithEmailAndPassword in useSignUp.ts in order to catch any errors and set the state back to 'ready'.